feat: [POC] merge collector and extractor
